+++ src/App.tsx
@@ -9,6 +9,8 @@ import LogoShowcase from './pages/LogoShowcase';
 import PrivacyPolicy from './pages/PrivacyPolicy';
 import TermsOfUse from './pages/TermsOfUse';
 import EmailPreview from './pages/EmailPreview';
+import DoNotSell from './pages/DoNotSell';
+import CookieBanner from './components/CookieBanner';
 
 function App() {
   // Use basename only in production (GitHub Pages)
@@ -28,7 +30,9 @@ function App() {
         <Route path="/scholarship/:scholarshipSlug" element={<ScholarshipPage />} />
         <Route path="/join/:parentId/:childId" element={<LandingPage />} />
         <Route path="/email-preview" element={<EmailPreview />} />
+        <Route path="/do-not-sell" element={<DoNotSell />} />
       </Routes>
+      <CookieBanner />
     </Router>
   );
 }